+++ src/br/com/erp/resources/CorImp.java
@@ -10,8 +10,11 @@ import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
 import br.com.erp.json.ParamJson;
+import br.com.erp.model.Contato;
 import br.com.erp.model.Cor;
 import br.com.erp.model.Marca;
+import br.com.erp.model.Produto;
+import br.com.erp.model.TipoContato;
 import br.com.erp.util.UnidadePersistencia;
 
 @Path("cor")
@@ -101,26 +104,58 @@ public class CorImp {
 		return cores;
 	}
 
+	
+	
 	@Path("removerPorId")
 	@POST
 	@Consumes(MediaType.APPLICATION_JSON + ";charset=utf-8")
 	@Produces(MediaType.APPLICATION_JSON + ";charset=utf-8")
 	public void remove(ParamJson paramJson) {
 		EntityManager em = UnidadePersistencia.createEntityManager();
-
+		
 		try {
 			Cor cor = em.find(Cor.class, paramJson.getInt1());
 			em.getTransaction().begin();
+			Produto produto = obterDependencia(paramJson.getInt1());
+			
+			if(produto != null) {
+				throw new RuntimeException("Existe uma dependência relacionada a este registro: " + produto.getId());
+			}
 			em.remove(cor);
 			em.getTransaction().commit();
 
 		} catch (Exception e) {
 			e.printStackTrace();
-			em.getTransaction().rollback();
-
+			if(em.getTransaction().isActive()){
+				em.getTransaction().rollback();
+			}
+			throw e;
 		} finally {
 			em.close();
 		}
 
 	}
+	
+	public Produto obterDependencia(Integer id) {
+		EntityManager em = UnidadePersistencia.createEntityManager();
+		
+		Produto produto = null;
+		
+		try {
+			produto = (Produto) em.createQuery("select a "
+											   +"from Produto a "
+											   +"where a.cor.id = :id")
+								  .setParameter("id", id)
+								  .setMaxResults(1)
+					              .getSingleResult();
+		}catch(Exception e){
+			e.printStackTrace();
+		}finally {
+			em.close();
+		}
+		
+		return produto;
+	}
+	
+	
 }
