+++ src/br/com/erp/resources/UnidadeMedidaImp.java
@@ -10,6 +10,8 @@ import javax.ws.rs.Path;
 import javax.ws.rs.Produces;
 import javax.ws.rs.core.MediaType;
 import br.com.erp.json.ParamJson;
+import br.com.erp.model.Marca;
+import br.com.erp.model.Produto;
 import br.com.erp.model.UnidadeMedida;
 import br.com.erp.util.UnidadePersistencia;
 
@@ -106,20 +108,50 @@ public class UnidadeMedidaImp {
 	@Produces(MediaType.APPLICATION_JSON + ";charset=utf-8")
 	public void remove(ParamJson paramJson) {
 		EntityManager em = UnidadePersistencia.createEntityManager();
-
+		
 		try {
 			UnidadeMedida unidadeMedida = em.find(UnidadeMedida.class, paramJson.getInt1());
 			em.getTransaction().begin();
+			Produto produto = obterDependencia(paramJson.getInt1());
+			
+			if(produto != null) {
+				throw new RuntimeException("Existe uma dependência relacionada a este registro: " + produto.getId());
+			}
 			em.remove(unidadeMedida);
 			em.getTransaction().commit();
 
 		} catch (Exception e) {
 			e.printStackTrace();
-			em.getTransaction().rollback();
-
+			if(em.getTransaction().isActive()){
+				em.getTransaction().rollback();
+			}
+			throw e;
 		} finally {
 			em.close();
 		}
 
 	}
+	
+	
+	public Produto obterDependencia(Integer id) {
+		EntityManager em = UnidadePersistencia.createEntityManager();
+		
+		Produto produto = null;
+		
+		try {
+			produto = (Produto) em.createQuery("select a "
+											   +"from Produto a "
+											   +"where a.unidadeMedida.id = :id")
+								  .setParameter("id", id)
+								  .setMaxResults(1)
+					              .getSingleResult();
+		}catch(Exception e){
+			e.printStackTrace();
+		}finally {
+			em.close();
+		}
+		
+		return produto;
+	}
+	
 }
